# Add [ThreadStatic] StringBuilder in JsonTextTokenizer