--- a/csharp/src/Google.Protobuf/JsonTokenizer.cs
+++ b/csharp/src/Google.Protobuf/JsonTokenizer.cs
@@ -177,12 +177,6 @@ namespace Google.Protobuf
             // The set of states in which a value is valid next token.
             private static readonly State ValueStates = State.ArrayStart | State.ArrayAfterComma | State.ObjectAfterColon | State.StartOfDocument;
 
-            private const int MAX_BUILDER_SIZE = 360;
-            private const int DEFAULT_CAPACITY = 16;
-
-            [ThreadStatic]
-            private static StringBuilder? t_cachedInstance;
-
             private readonly Stack<ContainerType> containerStack = new Stack<ContainerType>();
             private readonly PushBackReader reader;
             private State state;
@@ -307,7 +301,8 @@ namespace Google.Protobuf
             /// </summary>
             private string ReadString()
             {
-                var builder = AcquireStringBuilder();
+                //builder will not be released in case of an exception, but this is not a problem and we will create new on next Acquire
+                var builder = StringBuilderCache.Acquire();
                 bool haveHighSurrogate = false;
                 while (true)
                 {
@@ -322,7 +317,7 @@ namespace Google.Protobuf
                         {
                             throw reader.CreateException("Invalid use of surrogate pair code units");
                         }
-                        return GetStringAndRelease(builder);
+                        return StringBuilderCache.GetStringAndRelease(builder);
                     }
                     if (c == '\\')
                     {
@@ -414,7 +409,8 @@ namespace Google.Protobuf
 
             private double ReadNumber(char initialCharacter)
             {
-                var builder = AcquireStringBuilder();
+                //builder will not be released in case of an exception, but this is not a problem and we will create new on next Acquire
+                var builder = StringBuilderCache.Acquire();
                 if (initialCharacter == '-')
                 {
                     builder.Append("-");
@@ -443,7 +439,7 @@ namespace Google.Protobuf
                 }
 
                 // TODO: What exception should we throw if the value can't be represented as a double?
-                var builderValue = GetStringAndRelease(builder);
+                var builderValue = StringBuilderCache.GetStringAndRelease(builder);
                 try
                 {
                     double result = double.Parse(builderValue,
@@ -570,38 +566,6 @@ namespace Google.Protobuf
                     ContainerType.Document => State.ExpectedEndOfDocument,
                     _ => throw new InvalidOperationException("Unexpected container type: " + parent),
                 };
-            }
-
-            private static StringBuilder AcquireStringBuilder(int capacity = DEFAULT_CAPACITY)
-            {
-                if (capacity <= MAX_BUILDER_SIZE)
-                {
-                    var sb = t_cachedInstance;
-                    // Avoid stringbuilder block fragmentation by getting a new StringBuilder
-                    // when the requested size is larger than the current capacity
-                    if (capacity <= sb?.Capacity)
-                    {
-                        t_cachedInstance = null;
-                        sb.Clear();
-                        return sb;
-                    }
-                }
-                return new StringBuilder(capacity);
-            }
-
-            private static void Release(StringBuilder sb)
-            {
-                if (sb.Capacity <= MAX_BUILDER_SIZE)
-                {
-                    t_cachedInstance = sb;
-                }
-            }
-
-            private static string GetStringAndRelease(StringBuilder sb)
-            {
-                var result = sb.ToString();
-                Release(sb);
-                return result;
             }
 
             private enum ContainerType
@@ -765,6 +729,56 @@ namespace Google.Protobuf
                 {
                     // TODO: Keep track of and use the location.
                     return new InvalidJsonException(message);
+                }
+            }
+
+            /// <summary>Provide a cached reusable instance of stringbuilder per thread.</summary>
+            private static class StringBuilderCache
+            {
+                private const int MaxCachedStringBuilderSize = 360;
+                private const int DefaultStringBuilderCapacity = 16; // == StringBuilder.DefaultCapacity
+
+                [ThreadStatic]
+                private static StringBuilder? cachedInstance;
+
+                /// <summary>Get a StringBuilder for the specified capacity.</summary>
+                /// <remarks>If a StringBuilder of an appropriate size is cached, it will be returned and the cache emptied.</remarks>
+                public static StringBuilder Acquire(int capacity = DefaultStringBuilderCapacity)
+                {
+                    if (capacity <= MaxCachedStringBuilderSize)
+                    {
+                        StringBuilder? sb = cachedInstance;
+                        if (sb != null)
+                        {
+                            // Avoid stringbuilder block fragmentation by getting a new StringBuilder
+                            // when the requested size is larger than the current capacity
+                            if (capacity <= sb.Capacity)
+                            {
+                                cachedInstance = null;
+                                sb.Clear();
+                                return sb;
+                            }
+                        }
+                    }
+
+                    return new StringBuilder(capacity);
+                }
+
+                /// <summary>Place the specified builder in the cache if it is not too big.</summary>
+                private static void Release(StringBuilder sb)
+                {
+                    if (sb.Capacity <= MaxCachedStringBuilderSize)
+                    {
+                        cachedInstance = sb;
+                    }
+                }
+
+                /// <summary>ToString() the stringbuilder, Release it to the cache, and return the resulting string.</summary>
+                public static string GetStringAndRelease(StringBuilder sb)
+                {
+                    string result = sb.ToString();
+                    Release(sb);
+                    return result;
                 }
             }
         }