While profiling allocations in our code, I've found a lot of StringBuilder instances.
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/036079fa-c6ef-4a00-887d-061dc788c5d6)

Changes add similar functionality to the dotnet StringBuilderCache class in  JsonTextTokenizer.
They are not modifying existing logic and only will reduce relatively small text/number allocations by using cached instance.