@TrayanZapryanov 2024-02-11
Is it okay if I push another PR for changing JsonToken to readonly struct?
It is also quite high in memory allocations and is prepared to be immutable.