@jskeet 2024-02-11
> Is it okay if I push another PR for changing JsonToken to readonly struct?
> It is also quite high in memory allocations and is prepared to be immutable.

Certainly happy to see a PR - the fact that it's internal is helpful here.

(I'll try to review *this* PR some time next week. Looking *very* briefly though, I'd encourage you to consider ThreadLocal vs ThreadStatic; I was under the impression that these days ThreadLocal was generally preferred, but I could be wrong.)