@TrayanZapryanov 2024-02-11
@jskeet I've mimic code from here [StringBuilderCache](https://github.com/dotnet/runtime/blob/main/src/libraries/Common/src/System/Text/StringBuilderCache.cs) and it is using [ThreadStatic]. If you prefer I can try with ThreadLocal, but this code is "battle tested" in runtime and also in our software :).