@TrayanZapryanov 2024-02-12
> Another alternative to consider would be to have a private nested StringBuilderCache class instead of adding these methods directly to the JsonTextTokenizer - that we wouldn't need to add "StringBuilder" into all the names.

In general I also prefer this to be extracted as a general class and used in on every place with new StringBuilder().
If you like I can expand changes in this direction and apply all the other comments.