@jskeet 2024-02-12
> In general I also prefer this to be extracted as a general class and used in on every place with new StringBuilder().
> If you like I can expand changes in this direction and apply all the other comments.

I think it's worth extracting as a general class, but I'd *start* with it in JsonTextTokenizer rather than applying it to *all* uses of StringBuilder in the project. We can easily expand its usage further later.