@jskeet 2024-02-12
Right, I'll fetch and try to do some benchmarking when I can. It may not be this week though.