@jskeet 2024-02-13
Benchmark results:

Before optimizations:

| Method                    | Mean     | Error    | StdDev   | Gen0   | Gen1   | Allocated |
|-------------------------- |---------:|---------:|---------:|-------:|-------:|----------:|
| TestAllTypesProto3_Medium | 84.52 us | 1.639 us | 2.013 us | 8.0566 | 0.3662 |  49.39 KB |

After stringbuilder:

| Method                    | Mean     | Error    | StdDev   | Gen0   | Gen1   | Allocated |
|-------------------------- |---------:|---------:|---------:|-------:|-------:|----------:|
| TestAllTypesProto3_Medium | 78.62 us | 1.509 us | 2.066 us | 5.3711 | 0.2441 |  33.14 KB |

So it looks like this is probably worth doing (and the benefit will very much depend on the JSON being parsed).
The current code introduces new warnings though, which should be fixed.