@TrayanZapryanov 2024-02-14
@jskeet Where do you see "new warnings"?
As I cannot compile it - it is hard for me to see if compiler warns.
Also tests are somehow paused :
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/cb3754bd-8bf1-4ec9-b90c-55d76aa89057)

Maybe there is somewhere a check that should be set in order to be executed....(Check for Safety)