@TrayanZapryanov 2024-02-15
@JamesNK I've added note from where I've copied code.

Here it is a result with allocations once I've build this PS + JsonToken structure of the same benchmarked app.
![image](https://github.com/protocolbuffers/protobuf/assets/12857389/27fcc1b4-8529-4a13-8722-8a9efb9ef096)

In general it is 120MB less