@JamesNK 2024-02-15
I mean the steady state memory usage of the app will be higher. Cached StringBuilders stick around.