@jskeet 2024-03-01
> @jskeet Do you need something more from my side? Or this is enough to be merged?

Just time to get round to final checks and things. Sorry - I'm really busy and Protobuf is only one small part of my work, so it's hard to get round to things.