@TrayanZapryanov 2024-03-06
Same benchmark as in [here](https://github.com/protocolbuffers/protobuf/pull/15795)

main:
// * Summary *

BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3235/23H2/2023Update/SunValley3)
11th Gen Intel Core i9-11900K 3.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.200
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI

Job=DefaultJob  Error=3.84 ns  StdDev=3.21 ns

| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     281.1 ns |   1.01 KB |
| Big    | 256,221.4 ns |  642.7 KB |

PR:
Stringbuilder cache

| Method | Mean         | Allocated |
|------- |-------------:|----------:|
| Small  |     292.6 ns |     512 B |
| Big    | 231,857.0 ns |  260792 B |