@jskeet 2024-03-06
@mkruskal-google: This has been reviewed from a C# standpoint, and I've just marked it safe to run tests again. If you could review it to kick off the merge process, that'd be great.